This is a Job Searching Application that can help users search for their required job(s).
This is an app that can help user to search for jobs based on their requirements. There is a search bar where users can search for jobs based on their programming languages on which they wish to apply. There are 3 tabs :-
1. Full time Jobs - Here users can find jobs based on full time basis .
2. Part-Time Jobs - here users can find jobs based on part - time basis.
3. Contractual Basis - Here users can find jobs based on contractual basis.

Clicking on jobs from the above tabs or categories will redirect the user to the available jobs. Clicking on a particular job will bring the user to that jobs details like Qualifications , Responsibilities etc. If these match with users skill sets then there is a button that the user can press to apply for the job . This will redirect the user to the company's website .

All of this is done using an API Call from RapidAPI services. The name of the API is jsearch.

This app best works with light mode , it needs some further works in dark mode for complete functionability.